+++ crates/forge/src/main.rs
@@ -1,6 +1,6 @@
 use anyhow::{anyhow, bail, Context, Result};
 use camino::Utf8PathBuf;
-use clap::{Parser, Subcommand};
+use clap::{Parser, Subcommand, ValueEnum};
 use forge::scarb::config_from_scarb_for_package;
 use include_dir::{include_dir, Dir};
 use scarb_artifacts::{
@@ -49,6 +49,13 @@ enum ForgeSubcommand {
     CleanCache {},
 }
 
+#[derive(ValueEnum, Debug, Clone)]
+enum ColorOption {
+    Auto,
+    Always,
+    Never,
+}
+
 #[derive(Parser, Debug)]
 #[allow(clippy::struct_excessive_bools)]
 struct TestArgs {
@@ -78,6 +85,10 @@ struct TestArgs {
     /// Run all tests regardless of `#[ignore]` attribute
     #[arg(long, conflicts_with = "only_ignored")]
     include_ignored: bool,
+
+    /// Control when colored output is used
+    #[arg(value_enum, long, default_value_t = ColorOption::Auto, value_name="WHEN")]
+    color: ColorOption,
 }
 
 fn validate_fuzzer_runs_value(val: &str) -> Result<u32> {
@@ -117,6 +128,12 @@ fn extract_failed_tests(tests_summaries: Vec<TestCrateSummary>) -> Vec<TestCaseS
 }
 
 fn test_workspace(args: TestArgs) -> Result<bool> {
+    match args.color {
+        ColorOption::Always => env::set_var("CLICOLOR_FORCE", "1"),
+        ColorOption::Never => env::set_var("CLICOLOR", "0"),
+        ColorOption::Auto => (),
+    }
+
     let scarb_metadata = MetadataCommand::new().inherit_stderr().exec()?;
     let workspace_root = scarb_metadata.workspace.root.clone();